denoising/requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
cffi==1.17.1
filelock==3.18.0
fsspec==2025.5.1
grpcio==1.73.1
hydra-core==1.3.2
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==25.0
pandas==2.3.1
protobuf==6.31.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
